fix internal plugin's enabled flag ignore user plugin's enabled
- make `get-user-plugins-config` function as independent function

- when you load enabled plugins, use `get-user-plugins-config` function so that it recognizes user's plugin config per environment
